fix: remove index.js and get items on-the-fly, remove extraOverrides
